refactor(api): move usecase
send-email-for-account-recovery
